requirements.txt
streamlit
snowflake-connector-python
snowflake-sqlalchemy
python-dotenv
pandas
faker
pyarrow
SQLAlchemy
PyMySQL
fastavro